NCG-5940: added sound improvements

+++ modules/audio_device/audio_device_generic.h
@@ -24,7 +24,7 @@ class AudioDeviceGeneric {
   // MARK: Nicegram NCG-5828 call recording
   using CompletionRecorderCallback = std::function<void(const std::string& outputFilePath,
                                                         double durationInSeconds,
-                                                        std::vector<int16_t> rawData)>;
+                                                        std::vector<uint8_t> rawData)>;
   //
   // For use with UMA logging. Must be kept in sync with histograms.xml in
   // Chrome, located at
